test_input/photo_cutout_tool/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
rembg==2.0.50
pillow==10.1.0
numpy==1.26.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-core==2.14.1
pydantic-settings==2.1.0
typing-extensions==4.8.0
starlette==0.27.0
anyio==3.7.1
click==8.1.7
h11==0.14.0
idna==3.6
sniffio==1.3.0
exceptiongroup==1.2.0
annotated-types==0.6.0
certifi==2023.11.17
charset-normalizer==3.3.2
requests==2.31.0
urllib3==2.1.0
opencv-python==4.8.1.78
scikit-image==0.22.0
imageio==2.33.1
tifffile==2023.12.9
networkx==3.2.1
lazy-loader==0.3
packaging==23.2
scipy==1.11.4
pooch==1.8.0
cloudpickle==3.0.0
dask==2023.12.1
PyWavelets==1.5.0
onnxruntime==1.16.3
protobuf==4.25.1
coloredlogs==15.0.1
flatbuffers==23.5.26
humanfriendly==10.0
mpmath==1.3.0
sympy==1.12
torch==2.1.2
torchvision==0.16.2
filelock==3.13.1
fsspec==2023.12.2
jinja2==3.1.2
markupsafe==2.1.3